Refactor to make things more interactive
Adds the symfony console component along with an export command, and
refactors the proof of concept slightly so that it's easier to add other
types of forum, or "source", in future.

Still a lot of work to do including getting posts from a topic that
spans multiple pages; caching; timestamps... Not to mention importing.

All in good time!