Add auto-increment logic to Cache creation
If we run the exporter multiple times, it can get tedious constantly
having to `rm -rf cache/source-dir` to avoid cases where PHP insists
on throwing an error for directories that exist.

We could remove the directory from the script before attempting to
create the new one, but that's even more tedious, and it could be
useful having a copy of the previous export run saved locally.