Get started on the Import command
It's going to be messy to start with, such is the nature of lazily
dumping all the things in one big file. So far it will connect to
Flarum, grab the forum title to confirm it's the right one, then
prompt the user to choose from a list of tags or create a one.

All this stuff will be much better off in its own file somewhere,
but we can worry about that later when it actually imports posts.