Add user import handling
Adds a whole heap of stuff to loop through cached users and make sure
they exist in Flarum before it attempts to create any posts or
discussions linked to a potentially non-existant account.

There's a potential flaw in the catch block where, if a user exists but
for some reason isn't in `$this->users`, it'll fail. The chances of that
actually happening though are slim, unless someone registers online
while the import command is running. Food for thought, perhaps.