Add discussions
Also adds minimal functionality for getting posts from cache so
the content can be injected as the start_post for the discussion.

Currently creates discussions with title and content, although the
timestamps can't be set just yet due to issue #2 blocked upstream.

Also todo is getting the discussions to show as the true author,
which will likely need a slight rework to $this->api so it caches
and allows us to set the userId part of the token dynamically.