Return once, reset early, add debug line
This makes sure the API is *always* reset. as some cases would still
result in the recursion issue we were seeing before.

Also outputs the error detail in case data is invalid but it's not the
username or email that's the issue. We can tweak this later to actually
be able to handle invalid usernames and hopefully solve issue #3.