Add users to cache from the create method instead
If we only add new users to the cache when we have the final result
from createUser() then there's a good chance it'll create other users
before it actually returns anything to importUsers().

That's a recipe for disaster. By moving it here, we can pretty much
guarantee that we'll know about any user as long as we created it
or it already existed before the import script was actually run.