Improve user error handling and output
Completely revamps error handling when user creation throws an exception
of some sort. We'll still rethrow if it's not a 422, but any validation
errors that we're not specifically looking for will now fail gracefully.

It's also safe to assume that a user with matching email AND username is
probably the exact same person, if not one we created ourselves. To save
a lot of c<ENTER>c<ENTER>c<ENTER> spam (and probably accidental confirms
too) we now automatically use any identical accounts that already exist.