Fix #3 by autohandling non-alphadash usernames
Takes the regex from Illuminate\Validation\Concerns\ValidatesAttributes
and modifies it slightly to replace any non-alphadash strings with a
dash. If a username contains a space it'll trigger invalid email as
well, and that blocks us since it isn't handled. To avoid those
cases we'll just replace spaces in the email from the start.