Convert post content to Markdown
Ideally this needs to be done in the export process, possibly with the
service defining what format the post content will be (html, plain, md)
so that we can act accordingly when writing the files from the Cache
class. But, I'm lazy, and that means exporting everything again.

This uses PHP League's converter which for the most part seems to work
great, with the exception that it wraps plain links in angle brackets.
To get around that, we have our own LinkConverter that removes them.