Add creation times to discussions
This should really be `dateTime` to match the attributes that get
returned in the response, but it will only pass through to the
`PostReply` handler correctly if it's set as `time`.

No additional core hacks needed for this to work, since Flarum does it
on its own in the `setStartPost()` method after creating discussions.