Tags are fully working at last! Fixes #5
Adds the option to set the parent tag when creating a new one as part of
the import process. The build function will has a bit of added juice to
enable skipping of child tags, since Flarum is capped it two levels.

Coupled with a small update to pass the parent id along with the payload
and hey presto, we're in business! Tags seem to be in correct order too.