Set user join dates during creation. Closes #2
The one downside here is that we don't have exact times from the old
support forum on octobercms.com, but this doesn't matter too much as
Flarum's timestamps are also relative for the user profiles anyway.

Aside from an absolute ton of cleanup, Flaportum is good to go!
For my uses, at least.